Update for domain restriction just past scri+ and for batch usage of files on Vulcan.
The domain restriction is currently only working for IntervalZero. The grid will now only cut off a points to the right of scri and will not cutt of points on scri. The same applies for points on 1-t*kappap except that the cut off must be to the right.

The batch processing has ment a rewrite of the HDF data processing files. These now run as stand alone files. They can be treated as bash scripts use the -h option to get helpful instructions on their use.