Bug fixes.
There were issues with dt being calculated differently on different subdomains (numerical error), with incorrect boundary conditions being used for calculation of when the boundary portion of the operator should be applied, and the incorrect moment of communication in the solver. All fixed, testing to 18 decimals (round off error) on 1, 2 and 3 processors.