PT mpi method implemented.
Note that the code is not polished and older files, such as OneDAdvection.py may have some issues. A huge range of changes have been made that, I think, simplify the code. I have focussed, where possible on removing the dependencies between classes. Some notable dependencies are still in the code. The obvious example is the implementation of boundary conditions.