Fixed off by one bug.
Spinsfast py takes a variable lmax. This is the maxium lmax that is used in computation, not the l value that bounds all other variables. The consequence is that if you are iterating over all l values you must use range(lmax + 1). This has caused issues in the transformations, salm, and ghp.py