"Altered on several accounts:
1) additional debuging information

2) now includes the assumption that all domain dependences occur in the axis
after the first. So that, for example if the domain is 1-D, but the data
has shape (5, 101, 6) then the `domain dependency' occurs in the second axis.
If the domain was 2-D and the data had shape (5, 101, 51) then
the `domain dependency' would occur in the second and third axis.

Changes to account for this have occured throughout the file.

3) small changes to reduce work load for dest and source <0 but not equal to
-2"